move novels to novels-list
