+++ frontend/app/novels/novels-list/data-table.tsx
@@ -37,7 +37,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
 
-import { Button } from "../../components/ui/button"
+import { Button } from "../../../components/ui/button"
 import { Input } from "@/components/ui/input"
 
 import { NovelEntry } from "./novel-types"
