+++ frontend/app/novels/novels-list/input-cell.tsx
